requirements.txt
cat > requirements.txt << 'EOF'
# Editorial Summarization System - Dependencies

# Core ML/NLP
torch>=2.0.0
transformers>=4.35.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# OCR (optional - only for dataset creation)
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Evaluation
rouge-score>=0.1.2

# Web app
streamlit>=1.28.0

# Utilities
pathlib
EOF